Separate HelloController into AssignmentController and HelloController
